ejemplos/ejemplo-agente-con-memoria-sqlite/pyproject.toml
[project]
name = "ejemplo-agente-con-memoria-sqlite"
version = "0.1.0"
description = "Ejemplo de agente LangGraph con memoria a largo plazo usando SQLite (SqliteSaver)"
dependencies = [
    "langgraph",
    "langgraph-checkpoint-sqlite",
    "langchain-openai",
    "langchain-core",
    "python-dotenv"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
